% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.studies.R
\name{case.studies}
\alias{case.studies}
\title{This function creates the functions needed to run the various case studies.}
\usage{
case.studies(which, nsample = 500)
}
\arguments{
\item{which}{name of the case study.}

\item{nsample}{=500, sample size.}
}
\value{
a list of functions
}
\description{
This function creates the functions needed to run the various case studies.
}
