% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Cpporder}
\alias{Cpporder}
\title{sort vector y by values in vector x}
\usage{
Cpporder(y, x)
}
\arguments{
\item{y}{numeric vector}

\item{x}{numeric vector}
}
\value{
numeric vector
}
\description{
sort vector y by values in vector x
}
\keyword{internal}
