% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gof_adj_C1}
\alias{gof_adj_C1}
\title{helper functions to do p value adjustment}
\usage{
gof_adj_C1(dta, rnull, vals, TS, typeTS, TSextra, B = 1000L)
}
\arguments{
\item{dta}{data set}

\item{rnull}{R function (generate data under null hypothesis)}

\item{vals}{values of discrete random variable}

\item{TS}{function to calculate test statistics}

\item{typeTS}{integer indicating type of test statistic}

\item{TSextra}{list to pass to TS}

\item{B}{=1000 Number of simulation runs}
}
\value{
A matrix of powers
}
\description{
helper functions to do p value adjustment
}
\keyword{internal}
