% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gof_cont}
\alias{gof_cont}
\title{run gof tests for continuous data}
\usage{
gof_cont(x, pnull, rnull, phat, TS, typeTS, TSextra, B = 5000L)
}
\arguments{
\item{x}{A numeric vector of data}

\item{pnull}{R function (cdf)}

\item{rnull}{R function (generate data under null hypothesis)}

\item{phat}{function to set or estimate parameters of pnull}

\item{TS}{function that calculates test statistics}

\item{typeTS}{integer indicating type of test statistic}

\item{TSextra}{list to pass to TS}

\item{B}{(=5000) Number of simulation runs}
}
\value{
A matrix of numbers
}
\description{
run gof tests for continuous data
}
\keyword{internal}
