% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bins_cont.R
\name{make_bins_cont}
\alias{make_bins_cont}
\title{This function creates several type of bins for continuous data}
\usage{
make_bins_cont(
  x,
  pnull,
  qnull = NA,
  phat = function(x) -99,
  DataBased = FALSE,
  nbins = c(50, 10),
  minexpcount = 5,
  Range = c(-99999, 99999)
)
}
\arguments{
\item{x}{data set}

\item{pnull}{cdf under the null hypothesis}

\item{qnull}{=NA quantile function, if available}

\item{phat}{=function(x) -99  parameters for pnull}

\item{DataBased}{=FALSE bins based on data, not expected counts}

\item{nbins}{=c(50, 10) number of bins}

\item{minexpcount}{=5 smallest expected count per bin}

\item{Range}{=c(-99999, 99999) limits of possible observations, if any}
}
\value{
A list of bins and bin probabilities
}
\description{
This function creates several type of bins for continuous data
}
