% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bins_disc.R
\name{make_bins_disc}
\alias{make_bins_disc}
\title{This function creates several types of bins for discrete data}
\usage{
make_bins_disc(
  x,
  pnull,
  phat = function(x) -99,
  nbins = c(50, 10),
  minexpcount = 5
)
}
\arguments{
\item{x}{counts}

\item{pnull}{cumulative distribution function}

\item{phat}{=function(x) -99, function to estimated parameters, or -99}

\item{nbins}{=c(50, 10) number of bins}

\item{minexpcount}{=5 smallest expected count per bin}
}
\value{
A list of indices
}
\description{
This function creates several types of bins for discrete data
}
