% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{newTSdisc}
\alias{newTSdisc}
\title{a local function needed for the vignette}
\usage{
newTSdisc(x, pnull, param, vals)
}
\arguments{
\item{x}{An integer vector.}

\item{pnull}{cdf.}

\item{param}{parameters for pnull in case of parameter estimation.}

\item{vals}{A numeric vector with the values of the discrete rv.}
}
\value{
A vector with test statistics
}
\description{
a local function needed for the vignette
}
