% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_power.R
\name{plot_power}
\alias{plot_power}
\title{This function draws the power graph, with curves sorted by the mean power and smoothed for easier reading.}
\usage{
plot_power(pwr, xname = " ", title, Smooth = TRUE, span = 0.25)
}
\arguments{
\item{pwr}{a matrix of power values, usually from the twosample_power command}

\item{xname}{Name of variable on x axis}

\item{title}{(Optional) title of graph}

\item{Smooth}{=TRUE lines are smoothed for easier reading}

\item{span}{=0.25bandwidth of smoothing method}
}
\value{
plt, an object of class ggplot.
}
\description{
This function draws the power graph, with curves sorted by the mean power and smoothed for easier reading.
}
