% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{power_disc}
\alias{power_disc}
\title{find power of gof tests for continuous data}
\usage{
power_disc(
  pnull,
  rnull,
  vals,
  ralt,
  param_alt,
  phat,
  TS,
  typeTS,
  TSextra,
  B = 1000L
)
}
\arguments{
\item{pnull}{R function (cdf)}

\item{rnull}{R function (generate data under null hypothesis)}

\item{vals}{values of discrete distribution}

\item{ralt}{R function to generate data under alternative}

\item{param_alt}{parameters of ralt}

\item{phat}{function to estimate parameters from the data}

\item{TS}{function to calculate test statistics}

\item{typeTS}{integer indicating type of test statistic}

\item{TSextra}{list to pass to TS}

\item{B}{=1000 Number of simulation runs}
}
\value{
A matrix of powers
}
\description{
find power of gof tests for continuous data
}
\keyword{internal}
