% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rgof-package.R
\docType{package}
\name{Rgof-package}
\alias{Rgof}
\alias{Rgof-package}
\title{Rgof: 1d Goodness of Fit Tests}
\description{
Routines that allow the user to run a large number of goodness-of-fit tests. It allows for data to be continuous or discrete. It includes routines to estimate the power of the tests and display them as a power graph. The routine run.studies allows a user to quickly study the power of a new method and how it compares to some of the standard ones.
}
\author{
\strong{Maintainer}: Wolfgang Rolke \email{wolfgang.rolke@upr.edu} (\href{https://orcid.org/0000-0002-3514-726X}{ORCID})

}
\keyword{internal}
