% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signif.digits.R
\name{signif.digits}
\alias{signif.digits}
\title{This function does some rounding to nice numbers}
\usage{
\method{signif}{digits}(x, d = 4)
}
\arguments{
\item{x}{a list of two vectors}

\item{d}{=4 number of digits to round to}
}
\value{
A list with rounded vectors
}
\description{
This function does some rounding to nice numbers
}
