% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_methods.R
\name{test_methods}
\alias{test_methods}
\title{This function checks whether the correct methods have been requested}
\usage{
test_methods(doMethods, Continuous, WithWeights)
}
\arguments{
\item{doMethods}{="all" Which methods should be included?}

\item{Continuous}{is data continuous}

\item{WithWeights}{with weights?}
}
\value{
TRUE or FALSE
}
\description{
This function checks whether the correct methods have been requested
}
\keyword{internal}
