% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timecheck.R
\name{timecheck}
\alias{timecheck}
\title{estimate run time function}
\usage{
timecheck(x, pnull, phatx, wx, TS, typeTS, TSextra)
}
\arguments{
\item{x}{data set}

\item{pnull}{function to find cdf under  null hypothesis}

\item{phatx}{parameter estimates}

\item{wx}{vector of wights}

\item{TS}{test statistic}

\item{typeTS}{format of TS}

\item{TSextra}{additional info TS}
}
\value{
Mean computation time
}
\description{
estimate run time function
}
