% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ts_C}
\alias{ts_C}
\title{This function calculates the test statistics for continuous data}
\usage{
ts_C(typeTS, x, TS, pnull, param, w, qnull, TSextra)
}
\arguments{
\item{typeTS}{format of TS}

\item{x}{continuous data set}

\item{TS}{routine}

\item{pnull}{cdf under the null hypothesis}

\item{param}{estimated parameters}

\item{w}{routine to calculate weights}

\item{qnull}{quantile function}

\item{TSextra}{list passed to TS function}
}
\value{
A vector of numbers
}
\description{
This function calculates the test statistics for continuous data
}
\keyword{internal}
