% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TS_cont}
\alias{TS_cont}
\title{Find test statistics for continuous data}
\usage{
TS_cont(x, pnull, param, qnull)
}
\arguments{
\item{x}{A numeric vector.}

\item{pnull}{cdf.}

\item{param}{parameters for pnull  in case of parameter estimation.}

\item{qnull}{An R function, the quantile function under the null hypothesis.}
}
\value{
A numeric vector with test statistics
}
\description{
Find test statistics for continuous data
}
