% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ts_D}
\alias{ts_D}
\title{This function calculates the test statistics for discrete data}
\usage{
ts_D(typeTS, x, TS, pnull, param, vals, TSextra)
}
\arguments{
\item{typeTS}{format of TS}

\item{x}{discrete data set (counts)}

\item{TS}{routine}

\item{pnull}{cdf under the null hypothesis}

\item{param}{estimated parameters}

\item{vals}{values of discrete RV}

\item{TSextra}{list passed to TS function}
}
\value{
A vector of numbers
}
\description{
This function calculates the test statistics for discrete data
}
\keyword{internal}
