% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TSw_cont}
\alias{TSw_cont}
\title{Find test statistics for continuous data with weights}
\usage{
TSw_cont(x, pnull, param, w)
}
\arguments{
\item{x}{A numeric vector.}

\item{pnull}{cdf.}

\item{param}{parameters for pnull in case of parameter estimation.}

\item{w}{numeric vector of weights}
}
\value{
A numeric vector with test statistics
}
\description{
Find test statistics for continuous data with weights
}
\keyword{internal}
