% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TSw_disc}
\alias{TSw_disc}
\title{Find test statistics for discrete data}
\usage{
TSw_disc(x, pnull, param, vals, w)
}
\arguments{
\item{x}{An integer vector.}

\item{pnull}{cdf.}

\item{param}{parameters for pnull in case of parameter estimation.}

\item{vals}{A numeric vector with the values of the discrete rv.}

\item{w}{weights}
}
\value{
A vector with test statistics
}
\description{
Find test statistics for discrete data
}
\keyword{internal}
