% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wbincounter}
\alias{wbincounter}
\title{Find counts or sum of weights in bins. Useful for power calculations. Replaces hist command from R.}
\usage{
wbincounter(x, bins, w)
}
\arguments{
\item{x}{numeric vector}

\item{bins}{numeric vector}

\item{w}{numeric vector of weights}
}
\value{
sum of weights in bins
}
\description{
Find counts or sum of weights in bins. Useful for power calculations. Replaces hist command from R.
}
\keyword{internal}
