% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcTS}
\alias{calcTS}
\title{This function calculates the test statistics}
\usage{
calcTS(dta, pnull, param, TS, typeTS, TSextra)
}
\arguments{
\item{dta}{list with data set}

\item{pnull}{cdf under the null hypothesis}

\item{param}{estimated parameters (or some constant)}

\item{TS}{routine}

\item{typeTS}{format of TS}

\item{TSextra}{list passed to TS function}
}
\value{
A vector of numbers
}
\description{
This function calculates the test statistics
}
\keyword{internal}
