% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TS_disc}
\alias{TS_disc}
\title{Find test statistics for discrete data}
\usage{
TS_disc(x, Fx, vals)
}
\arguments{
\item{x}{An integer vector.}

\item{Fx}{A numeric vector of cumulative probabilities.}

\item{vals}{A numeric vector with the values of the discrete rv.}
}
\value{
A vector with test statistics
}
\description{
Find test statistics for discrete data
}
\keyword{internal}
